+++ win32/Strips.cxx
@@ -749,10 +749,11 @@ LRESULT BackgroundStrip::WndProc(UINT iMessage, WPARAM wParam, LPARAM lParam) {
 	return 0l;
 }
 
-void BackgroundStrip::SetProgress(const GUI::gui_string &explanation, int size, int progress) {
+void BackgroundStrip::SetProgress(const GUI::gui_string &explanation, size_t size, size_t progress) {
 	if (explanation != ControlGText(wExplanation)) {
 		::SetWindowTextW(HwndOf(wExplanation), explanation.c_str());
 	}
+	// TODO: Should scale values as PBM_SETRANGE32 limited to 32-bit
 	::SendMessage(HwndOf(wProgress), PBM_SETRANGE32, 0, size);
 	::SendMessage(HwndOf(wProgress), PBM_SETPOS, progress, 0);
 }
